+++ history.md
@@ -69,6 +69,7 @@ This release is largely API compatible, but makes several breaking changes.
 - Refactor URI parsing to happen earlier, in Request initialization.
 - When adding URL params, handle URLs that already contain params.
 - Multipart: use a much more robust multipart boundary with greater entropy.
+- Make `RestClient::Payload::Base#inspect` stop pretending to be a String.
 
 # 2.0.0.rc1